chore(server): update

+++ server/src/main/java/com/example/crudspringboot/controller/UserInfoController.java
@@ -1,14 +1,18 @@
 package com.example.crudspringboot.controller;
 
+import com.example.crudspringboot.model.Card;
 import com.example.crudspringboot.model.UserInfo;
 import com.example.crudspringboot.service.UserService;
 import com.example.crudspringboot.service.userInfo.UserInfoService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
 import java.util.List;
 
 @CrossOrigin
@@ -43,44 +47,61 @@ public class UserInfoController {
     }
 
     @PostMapping()
-    public Object createUser(@RequestBody UserInfo requestUser) {
+    public Object createUser(@RequestBody UserInfo requestUser, HttpServletRequest request) {
+        String encodedRole = request.getHeader("Role");
+        String decodedRole = new String(Base64.getDecoder().decode(encodedRole), StandardCharsets.UTF_8);
+        if ("ADMIN".equals(decodedRole)) {
+            if (userInfoService.existsByUsername(requestUser.getUsername())) {
+                return new ResponseStatusException(HttpStatus.BAD_REQUEST, "Пользователь существует");
+            }
 
-        if (userInfoService.existsByUsername(requestUser.getUsername())) {
-            return new ResponseStatusException(HttpStatus.BAD_REQUEST, "Пользователь существует");
+            UserInfo user = new UserInfo();
+            user.setUsername(requestUser.getUsername());
+            user.setPassword(passwordEncoder.encodePassword((requestUser.getPassword())));
+            user.setRoles(requestUser.getRoles());
+
+            UserInfo savedUser = userInfoService.addUserInfo(user);
+            return new ResponseEntity<>(savedUser, HttpStatus.CREATED);
+        } else {
+            return new ResponseStatusException(HttpStatus.BAD_REQUEST, "Ограничение прав доступа");
         }
-
-        UserInfo user = new UserInfo();
-        user.setUsername(requestUser.getUsername());
-        user.setPassword(passwordEncoder.encodePassword((requestUser.getPassword())));
-        user.setRoles(requestUser.getRoles());
-
-        UserInfo savedUser = userInfoService.addUserInfo(user);
-        return new ResponseEntity<>(savedUser, HttpStatus.CREATED);
 
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateUser(@PathVariable Long id, @RequestBody UserInfo user) {
-        UserInfo existingUser = userInfoService.findUserInfoById(id);
-        if (existingUser == null) {
-            return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND);
+    public Object updateUser(@PathVariable Long id, @RequestBody UserInfo user, HttpServletRequest request) {
+        String encodedRole = request.getHeader("Role");
+        String decodedRole = new String(Base64.getDecoder().decode(encodedRole), StandardCharsets.UTF_8);
+        if ("ADMIN".equals(decodedRole)) {
+            UserInfo existingUser = userInfoService.findUserInfoById(id);
+            if (existingUser == null) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND);
+            } else {
+                existingUser.setUsername(user.getUsername());
+                existingUser.setRoles(user.getRoles());
+                existingUser.setPassword(passwordEncoder.encodePassword((user.getPassword())));
+                userInfoService.addUserInfo(existingUser);
+                return new ResponseEntity<>("Данные пользователя успешно добавлены", HttpStatus.OK);
+            }
         } else {
-            existingUser.setUsername(user.getUsername());
-            existingUser.setRoles(user.getRoles());
-            existingUser.setPassword(passwordEncoder.encodePassword((user.getPassword())));
-            userInfoService.addUserInfo(existingUser);
-            return new ResponseEntity<>("Данные пользователя успешно добавлены", HttpStatus.OK);
+            return new ResponseStatusException(HttpStatus.BAD_REQUEST, "Ограничение прав доступа");
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteUserById(@PathVariable Long id) {
-        UserInfo user = userInfoService.findUserInfoById(id);
-        if (user == null) {
-            return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND);
+    public Object deleteUserById(@PathVariable Long id, HttpServletRequest request) {
+        String encodedRole = request.getHeader("Role");
+        String decodedRole = new String(Base64.getDecoder().decode(encodedRole), StandardCharsets.UTF_8);
+        if ("ADMIN".equals(decodedRole)) {
+            UserInfo user = userInfoService.findUserInfoById(id);
+            if (user == null) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND);
+            } else {
+                userInfoService.deleteUserInfo(user);
+                return new ResponseEntity<>("Пользователь успешно удален", HttpStatus.OK);
+            }
         } else {
-            userInfoService.deleteUserInfo(user);
-            return new ResponseEntity<>("Пользователь успешно удален", HttpStatus.OK);
+            return new ResponseStatusException(HttpStatus.BAD_REQUEST, "Ограничение прав доступа");
         }
     }
 }
